管理・ログイン/パスワードリセット周りのvite移行
