+++ app/views/admin/reset_passwords/edit.html.haml
@@ -1,4 +1,4 @@
-= javascript_pack_tag 'admin/reset_passwords/edit'
+= vite_pack_tag 'admin_reset_passwords_edit'
 
 %body.hold-transition.register-page
   .register-box
