+++ app/views/layouts/admin/_sidebar.html.haml
@@ -1,8 +1,8 @@
 / Main Sidebar Container
 %aside.main-sidebar.sidebar-dark-primary.elevation-4
   / Brand Logo
-  %a.brand-link{href: "index3.html"}
-    %img.brand-image.img-circle.elevation-3{alt: "AdminLTE Logo", src: "#{asset_pack_path 'media/images/AdminLTELogo.png'}", style: "opacity: .8"}/
+  %a.brand-link{href: admin_dashboards_path}
+    %img.brand-image.img-circle.elevation-3{alt: "AdminLTE Logo", src: "#{vite_asset_path 'images/AdminLTELogo.png'}", style: "opacity: .8"}/
     %span.brand-text.font-weight-light AdminLTE 3
   / Sidebar
   .sidebar
